Little Additions to 1.3 (clean)
added props parameter to move and animate move now calls the more
generic animate to save code.  Made the Pane and Waiter class work with
any container (not just the stage).  Adjusted the default of the swipe
distance and time to 50 pixels in 100 ms to improve swiping on mobile
and added these as properties so can set dynamically.  May still need to
experiment more - this was based on 8 different phones.  Exposed the
swipe object used in the ZIM Pages class as a property.  Documented the
items property on the Guide, Grid and Layout Manager classes.  Added a
type parameter to the scaleTo() function to handle smallest, largest and
both - for types of scaling when two scale directions provided.