updated zim.rand()
was returning a random number greater than upper bound when integer was
set to false - warning... if you adjusted for this, the behaviour has
changed - you will need to readjust.  Only affects when third parameter
was set to false (not the default).  Apologies.  If last parameter is
set to false, it returns a decimal number between a and b including a
but not including b.  If the last parameter is true (default) it returns
an integer between a and b and including a and b.