Update to Parallax class
ZIM Build Parallax class has a new signature and now allows for using
scrollX and scrollY to control parallax (as well as any custom input).
ZIM Code received a scrollX function and a bug fix to ProportionDamp for
rounding.  ZIM Create has its centerReg() provide a center on container
second parameter to center the reg and center on stage for example.
Added a stand alone sub module for parallax.