ZIM animate and move updates to rewind
Added convenience rewind, rewindWait, rewindCall and rewindParams to the
props parameter of the zim.move and zim.animate functions.  Fixed a bug
in centerReg so now works on objects with bounds other than x and y = 0
(for instance a zim.Circle - which had centered bounds already... so did
not notice).  Updated docs for zim.Parallax and zim.Triangle