ZIM 2.0 - Introducing DUO - two ways to do params
Created zob() function for ZIM Wrap that lets you pass in parameters the
old way or as a single object with parameters as properties.  For
example: function test(a,b,c){}; test(1,null,3); test({a:1,c:3}); Went
through and changed perhaps 60 ZIM functions and classes to use the DUO
technique.  So... all ZIM 2.0 should work backwards compatible and then
the new functionality sits on top for an extra 3K.  Primarily because
for the minified, the parameters need to be repeated as a string.