Update to zob - for constructors
Added a fourth parameter that is needed when calling function with new.
As in new MyFunction().  The last parameter you pass in the this
reference.  I could maybe just have made it a boolean for whether you
are calling a class.  If you do not pass in this then it will assume a
regular function call.