Adjust plain classes to pass "this" to zob
And fixed Parallax to not center x if auto is false.  This could mess up
if Parallax was used with custom slider... but at this point I doubt
anyone is doing that - I probably should have launched a minor change
but I invalidated the cloud instead... did I say the last time that the
last time would be the last time?