4.3.0 - added extend() and used on all Classes
introduced zim.extend(sub, super, override, prefix, prototype) which
piggybacks on createjs.extend and createjs.promote
now all extends are done using this so that zim classes can be properly
extended
this means that the createjs namespace must be loaded before zim - too
bad, but that's okay
all classes are now one function - had to add the scope parameter (this)
to zob() for these.
fixed memory leak in Button - was not clearing mouseout event - off()
sometimes does not work!
updated RadioButtons dot to default .7 alpha in color
fixed sequence to take into account the new loopCall and loopParams
had to adjust loop() so that a return is used to continue and a return
with a value is used to break
return "continue" no longer continues it actually breaks - the function
is only one version old so may as well fix it
updated the Button description and added pizzazz to icon param
description
added a rollPerSecond parameter to zim.Frame with a default of 20 checks
per second (was set at 10)