ZIM 4.7.2 - Dynamo, Accelerator, multiple Loading
frame.loadAssets() now returns a zim.Queue object where we can capture
events for a specific loading
Also there is an isLoading property on the Queue and reference to the
createjs.LoadQueue with a preload property
The frame also captures these events for all queues as before and now
has an overall isLoading property.
Thanks to Frank Los for prompting the changes and helping with the
solution.
loadAssets() has a new time parameter to force the preload to wait at
least this long in ms.
This can be used in testing and for when you want a minimum time to show
a preload message
Added a showing property to zim.Waiter() and zim.Pane()
You can now loop an animation inside an animation series forever - it
just will not proceed to the next animation
The set parameter for zim.move() and zim.animate() now lets you set
scale (the convenience property for scaleX and scaleY)
Went back two updates to fix the loopCall parameter which was not
calling when looping forever
since the change to the loopCall to not call on the last loop - use call
for that (nice sentence!)
Added a zim.Dynamo() class to handle dynamic Sprite animation
You pass the Dynamo a Sprite, a base frames per second an optional label
or start end frames
The Dynamo should be able to parse any type of CreateJS spritesheet data
including nested animation labels
It turns these into an array of frames animates through the frames with
percentSpeed available to adjust dynamically
The Dynamo also allows you to pause with a time delay or a on a frame
number - and of course, unpause
Added a zim.Accelerator() class to handle dynamically changing speed and
pausing of Dynamo and Scroller objects
This allows you to control the objects from one place to dynamically
change the speed of a scene
Upddated the Scroller class to allow for pausing and percentSpeed
And the Scroller no longer has to be added to the stage to work
and in either case, the scroller backgrounds do not have to be on the
stage to make a Scroller
See http://zimjs.com/code/zide/ and
http://zimjs.com/code/zide/index2.html for examples