ZIM 5 - zik, wiggle, Emitter, many updates
Introduced ZIM VEE values that let you submit options for parameters to
be chosen later (uses the new zik() function)
Added zik() global function to pick randomly from an Array or a function
result or a random number based on
an object with min, max, integer, negative properties like the params
for zim.rand()
Added zim.Emitter() to emit particles for fireworks, smoke, flames,
falling objects, hair, grass, etc.
This is a large class with 30 parameters some of which accept ZIM VEE
values
Examples can be found here: http://zimjs.com/code/particles
Added zim.wiggle() to randomly go back and forth between the value
passed to it - uses zik() and animate()
Adjusted zim.move and zim.animate to accept ZIM VEE values for many of
the parameters
Including obj, set and props properties and overall parameters except
for params, loopParams, waitParams, sequenceParams
These will then be filtered through zik() to pick a random values
This allows for a delay before a random value is picked - handy for
passing info as a parameter
Used by zim.Emitter() for instance to pass random properties and animate
settings
Added feature to zim.animate and zim.move to animate relative to current
property value
To do this the parameter value in the animation object is put in a
string - can include negative sign
Also works with the set parameter to set relative starting values
Fixed rewind with set parameter active - hopefully that works - gets
complicated
Fixed zim.Label used Number.isNaN which is not supported on mobile so
changed to older double test.
Fixed a bug in zim.copy - an undeclared var was flatenning nested
objects
Adjusted zim.copy to ignore objects that are not an object literal - now
will copy references to zim objects rather than breaking
Adjusted zim.Triangle() adjust parameter to move shape rather than
registration
as usually you are adjusting the centerReg and that just was negating
the adjustment
Added rotation to cloning of props - this was an oversight BREAK - but
in a good way
adjusted clone for shapes to clone latest color and border properties
(and width and height for Rectangle and radius for Circle) possible
BREAK
Added read only framerate property to docs for zim.Ticker - use setFPS()
to set framerate
Added a has(stage, function) static method to zim.Ticker to check if the
Ticker has that function for that stage
zim.move() and zim.animate() - added waitedCall, loopWaitCall and
rewindWaitCall to call functions after waiting and before loopWait and
rewindWait
zim.move() and zim.animate() - added waitedParams, waitParams,
loopWaitParams and rewindWaitParams to pass params to above (or
callbacks receive the target)
This BREAKS parameter order - sorry!  Hopefully by this time you are
using ZIM DUO configuration object technique
Adjusted rewindCall and loopCall to happen after rewindWait and loopWait
- so this may BREAK as it was before the waits!
Fixed a glitch in stopZimAnimate(ids) when removing animation with ids
with single animations across multiple targets - now works
Fixed a glitch in zim.interval where it was occasionally not pausing due
to race condition
Adjusted obj.clear to before the next() function in zim.timeout to
prevent errors on occasion
Added an zim.added() function and ZIM 4TH method to check if object has
been added to the stage
CreateJS has an "added" event that triggers when an object is added to
another container
but this container may not be on the stage.
added polls with a setInterval every 100ms to see if the object has a
stage property
Once it does then it calls the callback and removes the interval
Adjust scaleTo to take into account current scale - this could BREAK
code if you had accommodated this already
Adjusted zim.rand() with respect to a=0, b=0 explicit parameters
was returning Math.random() now returns 0.  Only no parameters return
Math.random() - might BREAK