5.3.1 - blendMode + more default stage changes
Added simple array of pages to Pages class if you do not need swipe data
for each page
Also, use the default stage as the default for the holder of Pages and
moved the holder parameter to the end
BREAKS - this breaks the parameter order of the Pages class
Since introducing the default stage, there are a few of these breaks.
BREAKS - Moved obj parameter for Grid and Guide to the last parameter as
it use the default stage
Added an optional index value for Pages go() so can use page object or
an index matching the pages array order
Swapped zim.Ticker.has(function, stage) as stage will default to default
stage BREAKS parameter order
Added a check to zim.Ticker.add() so that it will not add the same
function more than once ENHANCEMENT
added blendMode property to all zim Display objects that is the same as
compositeOperation
Fixed a glitch in Dial when min is greater than max - the ticks were not
showing (used Math.abs() for tickNumber)
BREAKS - moved the index to the second parameter of center and centerReg
and the add to the last parameter
as more often than not, we are using centerReg and center to add the
object to the stage and
the second most popular parameter is the level (index).  Sorry, this
breaks the parameter order for older apps