5.3.2 various small fixes and features + Multitouch
Adjusted zim.decimals() so that addZeros is the number of 0 spaces after
the decimal
addZeros is no longer a Boolean BREAKS
and a new parameter, addZerosBefore, is the number of 0 spaces before
the decimal
Also added a time parameter that just turns the decimal dot into a colon
to handle minutes and seconds
Added a total property to the ZIM intervalObject for the zim.interval()
that is -1 for infinite
Added a clear transition to Pages that fades out the old page then fades
in the new page each at half the transition time
Adjusted docs - the radius of a zim.Circle is its real radius before
scaling and setting the radius redraws the shape
as opposed to width and height which are not necessarily the bounds but
can be a scaled value
Fixed multitouch with ZIM - was using: if (touch)
createjs.Touch.enable(stage, true);
And yet unfortunately, the second parameter is true for singleTouch!
Ouch.
IMPROVEMENT So now using: createjs.Touch.enable(stage);
Adjusted zim.drag() so that drag Ticker is not removed if any touches
are still active
added transitioning property to pages so can test if pages are in
transition (as they are cached)