5.4.0 - gesture() for multitouch pan, pinch, rotate
Added zim.gesture() with pan, pinch and rotate multitouch gestures for
position, scale and rotation
zim.gesture() features a bounding rect, min and max scales, rotational
snapping and pan sliding.
IMPROVEMENT: This is a new method under the Methods module for all ZIM
Display objects
Also added associated zim.noGesture() and zim.gestureRect() Methods
for removing some or all of the gestures (and paradoxically, adding them
back)
and dynamically setting the bounding rectangle for the gestures.
Works with the touch parameter set to true for zim.Frame() which is the
default setting.