6.1.0 - Accessibility on Mobile + ResizeManager
6.1.0
Made zim.Accessibility work on iOS and mostly on Android - Stepper,
Slider, Dial and ColorPicker still need work
BREAK - changed applicationName to appName, and highlight to
alwaysHighlight and the parameters following are prefixed with AH
This is a solid week of programming for iOS Voice Over - visit on Slack
for learnings and advice
Still untested on stand-alone screen readers but good on Windows
Narrator, and the mobile readers.
Added currentValueEvent properties to Slider, Dial to get or set value
and dispatch event if set changes value
normally, if we are setting a value with code, we can do whatever we
would do in the event at that time
but with accessibility, on mobile, the options are being set by HTML and
we want to trigger the event
Added a text property to the RadioButton button containers
Added a zim.ResizeManager() that extends zim.Manager and handles
resizing Accessibility, Layout, Pages, Grid, Guide, TextArea and Loader
This is just a small class - a few lines of code.
Note - the Accesibility resize() will call any loader or text areas
inside it...
BREAKS Added decimals parameter to Accessibility before frame parameter
BREAKS for the following:
Adjusted Stepper type parameter to stepperType parameter to avoid
conflict with new type property for all Display Objects
Adjusted CheckBox type parameter to indicatorType parameter to avoid
conflict with new type property for all Display Objects
Adjusted Dial type parameter to indicatorType parameter to avoid
conflict with new type property for all Display Objects
Adjusted Indicator type parameter to indicatorType parameter to avoid
conflict with new type property for all Display Objects
BREAKS - changed buttons property of RadioButtons to an array rather
than container - to match buttons array in Tabs and Pad
Made points property of Blob read and write instead of just read
Added functionality for selectedIndex and stepperArray of Stepper to
work with "number" stepperType
Changed currentIndex to selectedIndex for Stepper - BREAKS - matches
property name with other components (sorry)
Updated docs for TextArea to include size parameter which was there all
along...
Adjusted Dial docs to read default 1 for step.