6.2.0 - Noise(), Tile(), + update to CreateJS 1.0
Updated ZIM to work with createjs.min.js the new CreateJS 1.0.0
There were a few minor depreciations to take care of (getStage() became
stage, setPaused() became pause)
These were like that to support versions older than 2015 but there is no
real use in that now.
Note the gpu parameter of Frame that defaults to false.
Set this to true to use StageGL which is the major part of the release.
See bubbling video when we were preparing for this:
https://www.youtube.com/watch?v=fUjyUpYdsM8
Added zim.Noise() class that makes OpenSimplex noise for art, terrains,
etc.
http://zimjs.com/code/noise/
This is part of the Code module along with Damp and Proportion, etc.
Added a zim.Tile() class to let you tile an object (or objects with a
ZIM VEE value)
You can choose set col and row parameters, add optional spacing and
mirror as you tile
This extends a zim.Container() and is part of the Display module
Added a zim.smoothStep(min, max, input) function to the code module
This takes an input value and outputs a value between 0 and 1
that represents a transition between the min and max with easing at both
ends
If you want the easing to be more pronounced, then reduce difference
between min and max
If the value falls outside the min or max it is set to the min or max
Remember the return value is between 0 and 1 so you can multiply by
max-min and add it to min
to get a value at the original scale
IMPROVEMENT added drawImageData() method and imageData property to
zim.Bitmap()
This allows you to draw pixel data into a Bitmap as you can with raw
Canvas
There is also reference to the proxyCanvas and the proxyContext for the
canvas the Bitmap uses as source
BREAKS - there are new width and height parameters for Bitmap that come
before the id parameter (breaks order)