6.2.2 - scale() fix, run() wait fix, red color
Added frame.red to colors (thanks Alexa!)
Fixed glitch in Sprite.run() with wait, waitedCall, waitedParams.
These were just being passed through to zim.animate() but were left all
as wait property names - oopsy.
Added pauseAnimate() and stopAnimate() methods (and functions) to
basically replace pauseZimAnimate and stopZimAnimate
This depreciates the last two - which were titled that early on as
dynamic methods before ZIM 4TH methods
All dynamic methods and properties had ZIM in them to avoid conflict
but now that methods are common, there is no need for ZIM in the method
name ;-).
Fixed scale() which was turned off for glitch with setMask
Traced it to a possible conflict with canvas scale() method only with
new CreateJS 1.0 and if other interactivity! - weird.
So scale() is back - all is well.