6.3.2 - Patch - swapped Parallax Params
Adjusted Parallax parameter order to layers, damp, auto, stage and give
the object the default frame's stage
With the scrollX, scrollY chrome fix, this is a good time to adjust the
stage parameter - it is the last of the ZIM classes left requiring a
stage since the default frame addition.
This was primarily due to parallax being its own module without Frame -
but now ZIM has Distill, so it is not that big a deal to loose parity
with the old stand-alone parallax module.