6.3.3 - scale() depreciated,  other fixes
BREAK the scale() method has been depreciated and removed - use sca()
which works the same way
This was the only case of something not working with the new CreateJS
version 1.
It only affected the Shape() class but still for consistency, we have
removed it.
There should not have been a problem but somehow obscuring scale would
break a CreateJS Shape
There does not appear to be a CreateJS scale method - but there is a
JavasScript Canvas scale method - so perhaps that is the culprit.
scale() was always a touch risky as a method name as it is so much like
a property so people were prone to ball.scale = 3; which does not work
BREAK - changed parameter order for TextArea and Loader - sorry - moved
frame parameter to end
We now have zimDefaultFrame which will do in most cases and if we are in
a second frame then pass that in to the frame paremeter
Added focus() property to TextArea  (thanks Armin for the prompting)
before was textArea.tag.focus() piggy-backing on JS focus of HTML tag
Adjusted OPTIMIZE, ANIMATE and DISTILL constants to work without the zim
namespace with zns false - accidentally made them work when zns was true
(so swapped)
IMPROVEMENT Fixed a glitch in run() method of Sprite() to allow for
replay of non-label Sprites
since the change to normalized playback supporting non-sequential frames
in SpriteSheets animations
Adjusted zim.place() to return object for chaining
IMPROVEMENT Adjusted zim.Rectangle to draw a normal rectangle if the
corner is 0 was drawing a rounded rectangle with corner 0
then realized the border was not mitred as expected - it had little
rounded corners instead of 90 degrees.