ZIM 6.4.0 - added Font to loadAssets + Pane change
IMPROVEMENT Added font support to zim.loadAssets() - for urls to fonts
and also Google fonts
Pass in font in format {font:name, src:url, type:string, weight:string,
style:string}
Any number of fonts can be passed in along with images, sound, etc. to
loadAssets()
The name is what you give it and you would use that name in a
zim.Label({font:"name"})
If you use http at the start of the font src it ignores the path
parameter for that font
This allows you to use your own fonts in a directory but also add google
fonts in the same loadAssets.
This piggy backs on CreateJS FontLoader() but consolidates the format
and combines with other assets to load
Added a backdropClose parameter to zim.Pane()
BREAK moved the container parameter to the end of zim.Pane as most panes
are added to the stage.
In the past, ZIM did not know the stage but now it has a ZimDefaultFrame
with its stage