4.4.1 - Transform and Blob updates!
IMPROVEMENT fixed memory leak in transform() that was leaving mousemove
events due to using wrong variable name
IMPROVEMENT BREAK removed dblclick parameter and replaced with toggle
parameter for single click editing
This is much more intuitive and has been set to a default of true.
So click on shows controls and click off hides controls.
IMPROVEMENT BREAK changed dblclick parameter for zim.Blob() to toggle to
match transform()
Also changed Blob to single click to show and hide controls.
IMPROVEMENT BREAK made blobs drag when controls are active to match
transform() and most editing tools
and changed dblclickDrag parameter to move to match transform() and
default move to true (set to false to turn off dragging)
IMPROVEMENT BREAK added a handleSize parameter to Blob just before the
toggle parameter that sets the handle size like with transform
This defaults to 20 on mobile and 10 on not mobile ;-)
BREAK changed the record() method of Blob to recordPoints() which
matches format when creating a Blob
IMPROVEMENT and then added new record() and set() methods that work
together for updating already created Blobs
This was so that Blob objects can be added to the zim.TransformManager
just objects with transforms
You will need to make a basic Blob with the correct number of points.
Also added a setPoints() method that sets the points of an existing Blob
that match the recordPoints
Strategy - if you want to create Blobs from data use the recordPoints
and pass that data into the points parameter.
If you want to adjust an existing Blob then use the setPoints() method
If you want to record more than just the point data - for instance, x,
y, color, etc. then use record() and set()
Or let the TransformManager handle it - it uses the new record() and
set() similar to the transformControls record() and set() methods.
Also adjusted zim.setMask() to help handle dragging of a Blob object as
a mask and added an example to the setMask docs
as well as an update to the snake at http://zimjs.com/code/z0oo.html
IMPROVEMENT - made zim.TransformManager also handle Blob objects as
mentioned above - including the saving of, resizing and handling show
and hide
Added getQueryString() method to return the HTML query string
?var=val&var2=val2 as an object with matching properties
Thanks Ami Hanya for the suggestion.
IMPROVEMENT - added loadTimeout parameter to frame.loadAssets() with a
default of 8000ms - this is how long to wait for assets
Thanks Jonghyun Kim for suggestion and to CreateJS for already having
that in the LoadQueue
IMPROVEMENT - adjusted frame.asset() to have a default "Broken Image"
object (ZIM Circles) when asset is not available.
This can be adjusted with the new loadFailObj parameter for zim.Frame()
- the object will have a type property of "EmptyAsset"
Also made loadAssets() accept a ZIM DUO configuration object.