6.6.0 - Squiggle, hitTestCircles, Button wait...
DOCS:
Removed namespace from documentation titles and doc examples
(remember that the namespace can still be used. It can also be required
by setting zns=true before importing ZIM)
Tidied up the function listings to not include title =
function(parameters) but rather title(parameters)
Added a text version of the docs for easy import into other viewing
systems like PDF (thanks Vishwas Gagrani)

SQUIGGLE:
Added ZIM Squiggle - like a blob but only a line
Squiggle has pretty well all the blob features including
TransformManager support
Squiggle does not have borderColor, borderWidth nor radius but rather
has color, thickness and length
It also defaults to "mirror" as the type of control rather than
"straight" which the Blob defaults to

BUTTON:
Added wait, waitTime, waitColor, rollWaitColor, waitTextColor,
rollWaitTextColor, waitModal, waitEnabled parameters to the end of the
Button parameters.
Added clearWait() and removeWait() methods, a waiting property and a
waited event which is dispatched if the waitTime is reached.
For wait backings and icons use the setBackings() and setIcons() methods
along with the waited event to handle changes.
BREAK - inserted a borderRollColor parameter before borderWidth which is
before corner.
Any existing buttons with traditional parameters up to corner will
broken - will have to insert an extra null before borderWidth.
Also added a borderRollColor property to the zim Button.
Since the parameter order has changed, moved dashed to after flatBottom
parameter - to match Rectangle.
Added wait parameters to Tabs and Pads as well

PANE:
Added the following parameters to ZIM Pane()
bar - (default null - no bar) a String or ZIM Label title for the pane
that will be presented on a bar across the top
barColor - (default "rgba(0,0,0,.2)") the background color of the bar if
a bar is requested
barHeight - (default fit label) the height of the bar if a bar is
requested
close - (default false) - a close X for the top right corner that closes
the pane when pressed
closeColor - (default #555) - the color of the close X if close is
requested
Also added bar, barColor, barBacking, and close properties to access
these objects
Fixed TextArea and Loader inside Pane add/removal problem - was using
Dictionary objects list rather than values list
Also for multiple TextAreas was not looping through container backwards
when removing so was missing an item
BREAK changed the backingAlpha to backdropColor to offer more
flexibility for the backdrop
use "rgba(0,0,0,.5)" etc. for dark and "rgba(256,256,256,.5)" for light
- and of course colors as well.
You can use a solid color and then drop the alpha with
pane.backdrop.alpha = .2 - not alp() as backing is CreateJS shape.

HITTESTCIRCLES:
Added ZIM hitTestCircles() method to all display objects.
This tests if two circle shapes are hitting using an equation based on
cirles made from bounds
If the bounds are not square, the circles are averaged - the radius is
half the average of the lengths of its sides
(This is different than hitTestCircle() where the points around the
"circle" actually stretch to an oval with the bounds)
obj.hitTestCircles(obj2) is a calculation based on intersecting circles
so is faster than shape hitTests
It will be as fast as hitTestBounds and hitTestGrid.
hitTestCircles second parameter is a margin number value which can be
positive or negative to tweak the hitTest
A margin parameter has been added to hitTestBounds as well before the
boundsShape

OTHER:
IMPROVEMENT - added zim.previewAudioSprite() to the META module to
create tabs with each sound in an AudioSprite for easy preview
Adjusted parseAudioSprite() to clarify that it returns a CreateJS
AudioSprite object
BREAK - adjusted Blob to have a setData() method rather than a set()
method so the general set() method is not obscured
BREAK - adjusted transform() transformObject to have a setData() method
rather than a set() method to be consistent with Blob and Squiggle
NOTE: transform has a transformObject which is a custom object - not a
Display object so set() is fine there but wanted to be consistent
BREAK - adjusted Blob and the transformObject to have recordData()
methods rather than record() methods
These now match - recordData(), setData(), recordPoints(), setPoints() -
sorry for the adjust but should be settled now.
Also adjusted the Docs on Blobs to update the single click edit.
Also adjusted the Docs on the Blob's points property to indicate a
different format for setting and getting points.
Adjusted TextArea and Loader to use zim.added() to test for the stage
and then on("added") to test for re-adding after
Fixed a glitch in Stepper so the downForward parameter works as expected
Added read only num property to Blob (to match Squiggle's num property)
for number of points
Added error support for font loading so load completes if font is
missing - takes the timeout though of 8 seconds
which is not ideal but could not seem to capture immediate bad URL error
with error or fileerror events
Fixed Stepper to set correct selectedIndex when stepping through number
list that does not start at 0 and make sure enabled has this number when
set to true
Gave Pane and Indicator enabled properties like most other components
Adjusted TextArea to have a background-color of rgba(0,0,0,.01) so
cursor and selection highlight shows up (was at transparent - weird that
would affect selection)
Adjusted zim.drag() so the last drag added is the only drag on the
object
Added a dynamically created audio tag for the first audio loaded to
jumpstart audio on some Apple devices
Thanks Bart Libert for reporting and testing the issue.