6.6.2 - added Ticker raw + Dynamo and Scroller fix
IMPROVEMENT Added raw() method to zim.Ticker() that directly adds a
function to a requestAnimationFrame
Also added a removeRaw() method to remove a raw function with an id made
by var id = Ticker.raw(function)
This gives ultimate speed without calling the library system of
Ticker.add()
Ticker.add() should still be used in general for multiple ticker calls
or calls along with zim animate, etc.
But use raw() for dynamic art for instance where you have one butter
smooth ticker.
You need to call stage.update() in the raw() function
We could have called stage.update() but it would run the risk of
duplicating the update if the coder also put it in
Updated Scroller to add second object to the same layer as the original
object
so no need to add Scroller objects to containers IMPROVEMENT
Added flip and flipVertical parameters to Dynamo to flip the Sprite if
speed is negative and reversible is false.
This was done by the coder previously - so it might BREAK a project
where it was already done... BREAK
Also changed reversable parameter to reversible for proper spelling ;-)
BREAK