ZIM 6.7.1 Added sha(), cur() - setMask update
Added chainable cache() method to Containers and Sprites - this
overrides CreateJS cache to operate on bounds by default and return the
object for chaining
so all Display Objects that extend a Container such as Circle,
Rectangle, Triangle, Blob, Squiggle, Label, etc. all get the new cache()
eg. var circle = new Circle(100, frame.red, frame.dark,
5).cache().drag();
cache() will automatically cache the dimensions of the Circle and add
the half borderWidth as well - NOTE that cache is chainable now!
You can still pass in dimensions and if you do, you can just pass the
first two parameters as width and height to start x and y at 0.
Or you can pass all four parameters like CreateJS cache() - note, the
fifth parameter is scale which can cache at a higher fidelity for later
scaling
This is a convenience function for circle.cache(-circle.radius,
-circle.radius, circle.radius*2, circle.radius*2);

Added a sha() chainable method for applying a createjs Shadow() that has
defaults or params for color, shiftX, shiftY, blur or a single CreateJS
Shadow
eg. var circle = new Circle(100, frame.red, frame.dark, 5).sha().drag();
This is a convenience function for : circle.shadow = new
createjs.Shadow("rgba(0,0,0,.3)", 5, 5, 10); // etc.

Updated gesture() and transform() to work properly with masking as well
as converted drag(), animate() and Blob() into new mask management
Added a dynamic parameter (default false) to setMask() that runs a
Ticker to track and update mask movement.
The Blob and any ZIM Shape that has drag, animate, transform or gesture
will set the dynamic property to true (unless it was explicitly set to
false)
A Shape() that is used for a mask does not have to be dynamic as it is
directly applied as the mask
If you are going to mask with a ZIM shape (Rectangle, Circle, Triangle)
and then manually transform the mask for instance, mask.sca(2);
then either setMask with dynamic of true or setMask after the
transformation
BREAK - IMPROVEMENT - made setMask() return object for chaining - now
that dynamic is set no need really to return mask object instead use
mask.zimMask
BREAK - Added count (total tiles) parameter to after cols and rows for
ZIM Tile - in case you do not want all columns filled on last row.
Added factor parameter to Swiper (default 1) is going the same direction
and -1 is going in opposite direction
Updated Swiper to recognize damp parameter
Added swipestop event to Swiper for when motion stops after swipeup
Adjusted added() method to clear longer interval if shorter interval
detects stage
Adjusted Blob and Squiggle to not check for stage updates when not on
stage due to removeFrom stage - was working fine for toggle but not if
manually removed
Added slidestop event to drag() docs