ZIM 6.7.2 Blob and Squiggle Points Updates
Updated Blob and Squiggle points and controls property.
Blob and Squiggle had their points parameter array set with Number
positions
The points property was initially read only for an array of point
objects like the control container, circle, and rectangles
This was fine until we introduced the TransformManager and added the
Blob and Squiggle.
At this time, we needed to be able to recreate blobs and squiggles.
We made points be able to set the points with an array in the same
format as the points parameter.
This left the points property with different formats for getting and
setting - not ideal.
So this change is fixing the issue and making point manipulation easier
and more consistent.
BREAK - the points property remains consistent in format for getting and
setting with x and y properties (and a control type):
points - [[controlX, controlY, circleX, circleY, rect1X, rect1Y, rect2X,
rect2Y, controlType], [etc]]
A new property has been added to get access to the zim objects for each
point:
pointObjects - [[control, circle, rect1, rect2, controlType], [etc.]]
Here, control is the container; circle, rect1 and rect2 are Circle and
Rectangle objects
Note, the order is now the same for consistency
We were also using the term "set" to represent the controls of one
point.
This starts getting confusing when we get and set values as we are using
the word set again. (Thanks Kris Gardiner for the advice)
So we now use control to refer to the controls at the point - control is
the container that holds the circle and the two squares
BREAK - the sets property has been changed to controls which provides
access to the overall container that holds all the control containers
BREAK - controls used to be a Boolean property to get or set the
visibility of the controls - this is now controlsVisible
Added an update parameter to Blob and Squiggle changeControl() -
defaults to false to let multiple changes be batched with update()
method