ZIM 6.7.5 amimation event, addTo, etc. defaults
BREAK ZIM move() has been removed.  Please use ZIM animate() instead
with obj:{x:val, y:val}
ZIM move() was a convenience method that just took an x and y parameter
and passed it through to animate().
Now with ZIM mov() it became too confusing to have both and move() was
not really needed.
IMPROVEMENT Added an events parameter to animate (default false) set to
true to dispatch "animation" events from the target
The "animation" event will be dispatched while the target is being
animated.  Thanks Ami Hanya for the recommendation.
Adjusted ZIM addTo(), center() and centerReg() to default to the adding
to the first frame's stage if no parameters
So now circle.addTo(), circle.center() or circle.centerReg() will add to
stage as long as there is one frame.
If there are two frames, it will get added to the first frame's stage so
beware.
BREAK The centerReg() used to just center the registration and not add
it to anything if there were no parameters.
Now, it centers and adds to the stage.  To just center the reg, use
centerReg(null, null, false);
It was quite rare to just center the reg and not add it to a container.
And it was important to keep the addTo(), center() and centerReg()
defaults consistent.
Added container and add parameters to ZIM pos()
For now, the container property defaults to null.
We will consider defaulting to the first frame's stage (if not already
added to a container) in the future.
The add parameter defaults to true and could be set to false to not add
to the container.
Thanks Kate Xue for discussion.