ZIM 7 - HEP - Keyboard, KeyFocus, and lots more
NEW
Added a Keyboard() class to the Display Objects (thanks Frank Los for
the design and code)
Keyboard works with Label() to provide single line editable text without
needing the system keyboard
Added top() and bot() chainable methods to put object at top or bottom
of container - thanks Ami Hanya for the recommendation.
IMPROVEMENT Added links to libraries at the top of the docs for SOCKET,
GAME, PHYSICS, THREE, PIZZAZZ, PIZZAZZ 2
as well as made all these files consistent in introduction and
documentation format (inside CDN files)
Launched Physics and Three as version 1.0 - Game and Socket were already
version 1 and Pizzazz is different
IMPROVEMENT Added video links to Docs - a little button near view and
bits at the bottom of each doc entry
IMPROVEMENT Added wheel event to Window for mousewheel scrolling - was
using out-of-date events
Added a scrolling event to Window when it is scrolled
Added an HTML NodeList to ZIM loop() - see ZIM zet() to get a NodeList
like $() in JQuery
Dedicated ZIM Portal to Stephen Hawkings - may he portal on!
Added Button borderColor property - thanks Chris Spolton!
Added an svgToBitmap() conversion function to META module to convert SVG
to a ZIM Bitmap
Thanks Kenil Domadia for the recommendation
For Base64 Bitmap support made Bitmap update the stage after 50ms and
then after another 50ms
this is because there is a delay in creating a Bitmap from data - thanks
Ami Hanya for the prompting

KEYFOCUS - IMPROVEMENT
Added a global KEYFOCUS constant to keep track of which component has
keyboard focus
This is a system for all components with keyboard controls (set to
active)
The first component made will set KEYFOCUS to itself
Anytime a component is used it sets KEYFOCUS to itself
Components have a keyFocus property that can be manually set
There is only one component with key focus
so setting removes key focus from the last key focused component
Added KEYFOCUS to TypeScript
If tabbing from one component to the next is needed, consider using ZIM
Accessibility()
Made keydown change components only if component is on stage

POS()
BREAK inserted an index parameter to pos(x,y,container,index,add) - also
now accepts DUO
Was going to make the default be adding to stage - but after some
thought, decided not to
You must still explicitly specify the container
We had changed it but then were getting errors where things were showing
up unexpectedly in score boards, etc
and it just seems a little too dangerous

ZIM GESTURE
Added onTop and surround parameters to ZIM gesture() - thanks Frank Los
for the suggestions
surround currently does not work if rotate is true
Adjusted gesture() rect parameter to handle resized and rotated objects
This is an IMPROVEMENT but also slightly BREAKS code with old rect
values
Here you specify the rectangle to contain the object - not the
registration point of the object
So this is slightly different than the rect provided to ZIM drag
To accommodate this we have added an update parameter to the gestureRect
This defaults to true and Gesture will dynamically update the
gestureRect as the object scales and rotates
Internally, it passes a false to the new update parameter.

GENERAL FIXES
Removed ColorPicker indicator if currentIndex is less than 0
Fixed grip on ColorPicker to be dragable by setting shape to
mouseEnabled false so background is selectable
IMPROVEMENT - adjusted TextArea so readOnly parameter works
Fixed decimal values for Stepper - some values had micro additions -
these have been rounded to the decimal number
Adjusted tabs so that if currentEnabled is true, clicking button leaves
button on rollcolor - like a button would - good for button bars
Fixed stage property in waitModal for buttons, tabs, etc.
Fixed next mousedown and hold on Stepper when mouse not moved.
Made animate() return the target for chaining when ANIMATE is set to
false.
Changed loop() docs for method to not include loop function
functionality
but rather refer to the loop() function in CODE module
BREAK - changed the Squiggle borderDashedCommand to dashedCommand as the
Squiggle has no border
Squiggle is no longer implements a ZIM Shape in TypeScript as it has no
border
Adjusted outline of Label to show above backing and backgroundColor
Adjusted wiggle to not use animate with a Ticker if object does not have
a stage
Adjusted Slider so inside setting properly centers the slider in bounds
Added JS document.Window and document.Blob references before overwriting
with ZIM Window and Blob